requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
lxml==5.4.0
lxml_html_clean==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
urllib3==2.4.0
